+++ datacapture/src/main/java/com/google/android/fhir/datacapture/MoreQuestionnaireItemComponents.kt
@@ -87,9 +87,6 @@ internal fun Questionnaire.QuestionnaireItemComponent.findVariableExpression(
   return variableExpressions.find { it.name == variableName }
 }
 
-internal const val CQF_CALCULATED_EXPRESSION_URL: String =
-  "http://hl7.org/fhir/StructureDefinition/cqf-calculatedValue"
-
 // Item control code, or null
 internal val Questionnaire.QuestionnaireItemComponent.itemControl: ItemControlTypes?
   get() {
