+++ datacapture/src/main/java/com/google/android/fhir/datacapture/validation/ValidationUtil.kt
@@ -16,27 +16,21 @@
 
 package com.google.android.fhir.datacapture.validation
 
-import ca.uhn.fhir.context.FhirContext
-import ca.uhn.fhir.context.FhirVersionEnum
-import ca.uhn.fhir.context.support.DefaultProfileValidationSupport
-import com.google.android.fhir.datacapture.CQF_CALCULATED_EXPRESSION_URL
-import org.hl7.fhir.r4.hapi.ctx.HapiWorkerContext
+import com.google.android.fhir.datacapture.enablement.fhirPathEngine
 import org.hl7.fhir.r4.model.Expression
 import org.hl7.fhir.r4.model.Type
-import org.hl7.fhir.r4.utils.FHIRPathEngine
+
+internal const val CQF_CALCULATED_EXPRESSION_URL: String =
+  "http://hl7.org/fhir/StructureDefinition/cqf-calculatedValue"
 
 fun Type.valueOrCalculateValue(): Type? {
   return if (this.hasExtension()) {
     this.extension.firstOrNull { it.url == CQF_CALCULATED_EXPRESSION_URL }?.let {
       val expression = (it.value as Expression).expression
-      fhirPathEngine.evaluate(this, expression).firstOrNull()?.let { it as Type }
+      fhirPathEngine.evaluate(this, expression).singleOrNull()?.let { it as Type }
     }
   } else {
     this
   }
 }
 
-private val fhirPathEngine: FHIRPathEngine =
-  with(FhirContext.forCached(FhirVersionEnum.R4)) {
-    FHIRPathEngine(HapiWorkerContext(this, DefaultProfileValidationSupport(this)))
-  }
